feat: Implement abstraction
Now a player can play more than 1 strategy at the same game
