+++ ConsoleApp/Program.cs
@@ -8,10 +8,10 @@ static class ConsoleApp
     public static void Main()
     {
         // Instanciar jugadores
-        CircularList<IPlayer> players = new CircularList<IPlayer>(new Dustin());
-        players.AddLast(new Dustin("Eleven"));
-        players.AddLast(new Dustin("Mike"));
-        players.AddLast(new Dustin("Lucas"));
+        CircularList<IPlayer> players = new CircularList<IPlayer>(new Dustin(new List<IStrategy>(){new BotaGorda()}));
+        players.AddLast(new Dustin("Eleven", new List<IStrategy>(){new BotaGorda()}));
+        players.AddLast(new Dustin("Mike", new List<IStrategy>(){new BotaGorda()}));
+        players.AddLast(new Dustin("Lucas", new List<IStrategy>(){new BotaGorda()}));
 
         // default
         int maxToken = 6;
