+++ DominoLibrary/Strategies/IStrategy.cs
@@ -1,0 +1,6 @@
+namespace DominoLibrary;
+
+public interface IStrategy
+{
+    Token_onBoard Play(PlayInfo info, IPlayer player);
+}
